# frontend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/surveys": {
      "post": {
        "operationId": "SurveyController_createSurvey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSurveyDtoRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}": {
      "patch": {
        "operationId": "SurveyController_updateSurvey",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSurveyDtoRequest"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      },
      "delete": {
        "operationId": "SurveyController_softDeleteSurvey",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      },
      "get": {
        "operationId": "SurveyController_getSurveyForView",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}/publish": {
      "post": {
        "operationId": "SurveyController_publishSurvey",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}/close": {
      "post": {
        "operationId": "SurveyController_closeSurvey",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}/duplicate": {
      "post": {
        "operationId": "SurveyController_duplicateSurvey",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/public": {
      "get": {
        "operationId": "SurveyController_searchPublicSurveys",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}/members": {
      "post": {
        "operationId": "SurveyController_addMember",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSurveyMemberDtoRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}/members/{userId}": {
      "delete": {
        "operationId": "SurveyController_removeMember",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/pages/{pageId}/questions": {
      "post": {
        "operationId": "SurveyController_createQuestion",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDtoRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      },
      "get": {
        "operationId": "SurveyController_getQuestions",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/questions/{questionId}": {
      "patch": {
        "operationId": "SurveyController_updateQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDtoRequest"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      },
      "delete": {
        "operationId": "SurveyController_deleteQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/pages/{pageId}/questions/reorder": {
      "post": {
        "operationId": "SurveyController_reorderQuestions",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderQuestionsDtoRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/questions/{questionId}/move": {
      "post": {
        "operationId": "SurveyController_moveQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveQuestionDtoRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}/structure": {
      "get": {
        "operationId": "SurveyController_getSurveyStructure",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Surveys"]
      }
    },
    "/surveys/{surveyId}/responses/start": {
      "post": {
        "operationId": "ResponseController_startResponse",
        "parameters": [
          {
            "name": "surveyId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Responses"]
      }
    },
    "/responses/{responseId}/answers": {
      "post": {
        "operationId": "ResponseController_saveAnswers",
        "parameters": [
          {
            "name": "responseId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SaveAnswersDtoRequest" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Responses"]
      }
    },
    "/responses/{responseId}/submit": {
      "post": {
        "operationId": "ResponseController_submitResponse",
        "parameters": [
          {
            "name": "responseId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Responses"]
      }
    },
    "/responses/{responseId}/reopen": {
      "post": {
        "operationId": "ResponseController_reopenResponse",
        "parameters": [
          {
            "name": "responseId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Responses"]
      }
    },
    "/responses/{responseId}": {
      "delete": {
        "operationId": "ResponseController_softDeleteResponse",
        "parameters": [
          {
            "name": "responseId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "tags": ["Responses"]
      }
    },
...
  },
  "info": {
    "title": "Survix API",
    "description": "Backend API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "CreateSurveyDtoRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Customer Feedback" },
          "description": { "type": "string", "example": "Q1 2026" },
          "organizationId": {
            "type": "string",
            "example": "67bc7ed84799cfd093cc2903"
          },
          "visibility": {
            "type": "string",
            "enum": ["PUBLIC", "PRIVATE"],
            "example": "PUBLIC"
          },
          "allowAnonymous": { "type": "boolean", "example": true },
          "allowMultipleResponses": { "type": "boolean", "example": false },
          "startsAt": {
            "type": "string",
            "example": "2026-02-24T18:10:00.000Z"
          },
          "endsAt": { "type": "string", "example": "2026-03-01T18:10:00.000Z" },
          "randomizeQuestions": { "type": "boolean", "example": false }
        },
        "required": [
          "title",
          "visibility",
          "allowAnonymous",
          "allowMultipleResponses"
        ]
      },
      "UpdateSurveyDtoRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "visibility": { "type": "string", "enum": ["PUBLIC", "PRIVATE"] },
          "allowAnonymous": { "type": "boolean" },
          "startsAt": {
            "type": "string",
            "example": "2026-02-24T18:10:00.000Z"
          },
          "endsAt": { "type": "string", "example": "2026-03-01T18:10:00.000Z" },
          "randomizeQuestions": { "type": "boolean" }
        }
      },
      "AddSurveyMemberDtoRequest": {
        "type": "object",
        "properties": {
          "userId": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["OWNER", "EDITOR", "ANALYST", "VIEWER"]
          }
        },
        "required": ["userId", "role"]
      },
      "CreateQuestionDtoRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "SHORT_TEXT",
              "LONG_TEXT",
              "CHECKBOX",
              "RADIO",
              "RATING",
              "DATE",
              "FILE_UPLOAD",
              "AUDIO",
              "VIDEO",
              "RANKING"
            ]
          },
          "title": { "type": "string", "example": "How satisfied are you?" },
          "description": { "type": "string" },
          "isRequired": { "type": "boolean", "example": true },
          "settings": { "type": "object" }
        },
        "required": ["type", "title", "isRequired"]
      },
      "UpdateQuestionDtoRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "isRequired": { "type": "boolean" },
          "settings": { "type": "object" }
        }
      },
      "ReorderQuestionsDtoRequest": {
        "type": "object",
        "properties": {
          "orderedQuestionIds": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["orderedQuestionIds"]
      },
      "MoveQuestionDtoRequest": {
        "type": "object",
        "properties": { "targetPageId": { "type": "string" } },
        "required": ["targetPageId"]
      },
      "AnswerItemDto": {
        "type": "object",
        "properties": {
          "questionId": { "type": "string" },
          "value": { "type": "object" }
        },
        "required": ["questionId", "value"]
      },
      "SaveAnswersDtoRequest": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AnswerItemDto" }
          }
        },
        "required": ["answers"]
      }
    }
  }
}